requirements.txt
pygame~=2.5.2
numpy~=1.25.2
numba~=0.59.1
splines~=0.3.1
scipy~=1.13.0
pyFFTW~=0.13.1
pillow~=10.3.0